feat: moving to uuid lib and nr1-community usage for NerdGraphErrors.
