+++ package.json
@@ -13,6 +13,7 @@
     "eslint-fix": "eslint nerdlets/ --fix"
   },
   "dependencies": {
+    "@newrelic/nr1-community": "^1.1.1",
     "graphiql": "^0.17.0",
     "graphiql-explorer": "https://github.com/mattbaker/graphiql-explorer.git",
     "graphql": "^14.0.2",
@@ -21,7 +22,8 @@
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-json-tree": "^0.11.2",
-    "react-select": "^3.0.4"
+    "react-select": "^3.0.4",
+    "uuid": "^7.0.2"
   },
   "devDependencies": {
     "@semantic-release/changelog": "^3.0.5",
